split publish tf and publish bin info
