+++ jsk_2016_01_baxter_apc/launch/segmentation_in_bin.launch
@@ -17,6 +17,12 @@
     <rosparam file="$(arg LOWER_SHELF)" command="load" ns="lower_shelf"/>
   </node>
 
+  <!-- publish BinInfo.msg from json -->
+  <node pkg="jsk_apc2016_common" type="publish_bin_tf.py"
+    name="publish_bin_tf" output="screen">
+    <remap from="~input/bin_info_array" to="/set_bin_param/bin_array"/>
+  </node>
+
   <include file="$(find jsk_2016_01_baxter_apc)/launch/include/segmentation_for_each_hand.launch">
     <arg name="INPUT_IMAGE" value="left_softkinetic_camera/rgb/image_color" />
     <arg name="INPUT_CLOUD" value="left_softkinetic_camera/depth/points" />
